+++ cmeta/category_api_v1.py
@@ -21,11 +21,15 @@ class Category(InitCategory):
     Standard Base Category with artifact management functions
     """
 
-    def __init__(self, 
-                 cm = None,
-                 module_file_path = None,
-                 logger: logging.Logger = None):
-        super().__init__(cm, module_file_path, logger)
+#    def __init__(self, 
+#                 cm = None,
+#                 module_file_path = None,
+#                 logger: logging.Logger = None):
+#        super().__init__(cm, module_file_path, logger)
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, module_file_path = __file__, **kwargs)
+
     
     ############################################################
     def test(self,